# Improve performance of user-callbacked sort functions