@nielsdos 2025-03-28
I'm going to leave it at this for now. Improving further is possible by handling the deprecation logic slightly smarter but that only gives a few milliseconds improvement and does not seem worth it given that the code will be gone in the near future anyway.
So no changes since the review except for me adding a description to the PR with benchmarks and an explanation.